requirements.txt
tapo>=0.8.7
python-dotenv>=1.0.0
psutil>=5.9.0

